Updated double to decimal

+++ Business.Tests/Models/MonthlyAccountCategorySumsTests.cs
@@ -19,22 +19,22 @@ namespace Business.Tests.Dashboard
 
         private static readonly object[] FortuneAndInvestmentScenarios =
         [
-            new object[]{ new Dictionary<MonthAndYear, Dictionary<string, double>>(), new MonthAndYear(2019, 9), 0 },
-            new object[]{ new Dictionary<MonthAndYear, Dictionary<string, double>>
-                { {new MonthAndYear(2019,9), new Dictionary<string, double> { {"Fortune", 2000 } } } },
+            new object[]{ new Dictionary<MonthAndYear, Dictionary<string, decimal>>(), new MonthAndYear(2019, 9), 0 },
+            new object[]{ new Dictionary<MonthAndYear, Dictionary<string, decimal>>
+                { {new MonthAndYear(2019,9), new Dictionary<string, decimal> { {"Fortune", 2000 } } } },
                 new MonthAndYear(2019, 9), 2000
             },
-            new object[]{ new Dictionary<MonthAndYear, Dictionary<string, double>>
-                { {new MonthAndYear(2019,9), new Dictionary<string, double> { {"Investment", 3000 } } } },
+            new object[]{ new Dictionary<MonthAndYear, Dictionary<string, decimal>>
+                { {new MonthAndYear(2019,9), new Dictionary<string, decimal> { {"Investment", 3000 } } } },
                 new MonthAndYear(2019, 9), 3000
             },
-            new object[]{ new Dictionary<MonthAndYear, Dictionary<string, double>>
-                { {new MonthAndYear(2019,9), new Dictionary<string, double> { { "Fortune", 2000 }, { "Investment", 3000 } } } },
+            new object[]{ new Dictionary<MonthAndYear, Dictionary<string, decimal>>
+                { {new MonthAndYear(2019,9), new Dictionary<string, decimal> { { "Fortune", 2000 }, { "Investment", 3000 } } } },
                 new MonthAndYear(2019, 9), 5000
             },
         ];
         [TestCaseSource(nameof(FortuneAndInvestmentScenarios))]
-        public void GetFortuneAndInvestmentSummed_DifferentScenarios_ShouldReturnCorrectValue(Dictionary<MonthAndYear, Dictionary<string, double>> dictionary, MonthAndYear monthAndYear, double expectedSum)
+        public void GetFortuneAndInvestmentSummed_DifferentScenarios_ShouldReturnCorrectValue(Dictionary<MonthAndYear, Dictionary<string, decimal>> dictionary, MonthAndYear monthAndYear, decimal expectedSum)
         {
             var sut = new MonthlyAccountCategorySums(dictionary);
