+++ Business/Models/AccountStatus.cs
@@ -2,11 +2,11 @@
 
 namespace Business.Models
 {
-    public class AccountStatus(int accountStatusId, int accountId, DateTime date, double amount)
+    public class AccountStatus(int accountStatusId, int accountId, DateTime date, decimal amount)
     {
         public int AccountStatusId { get; private set; } = accountStatusId;
         public int AccountId { get; private set; } = accountId;
         public DateTime Date { get; private set; } = date;
-        public double Amount { get; private set; } = amount;
+        public decimal Amount { get; private set; } = amount;
     }
 }
