+++ Business/Models/Ledger.cs
@@ -2,13 +2,13 @@
 {
     public class Ledger
     {
-        public double Income { get; set; } = 0;
-        public double Expenses { get; set; } = 0;
-        public double VariableExpenses { get; set; } = 0;
-        public double FixedExpenses { get; set; } = 0;
-        public double Fortune { get; set; } = 0;
-        public double InvestedFortune { get; set; } = 0;
-        public double FortuneIncrease { get; set; } = 0;
-        public double InvestmentIncrease { get; set; } = 0;
+        public decimal Income { get; set; } = 0;
+        public decimal Expenses { get; set; } = 0;
+        public decimal VariableExpenses { get; set; } = 0;
+        public decimal FixedExpenses { get; set; } = 0;
+        public decimal Fortune { get; set; } = 0;
+        public decimal InvestedFortune { get; set; } = 0;
+        public decimal FortuneIncrease { get; set; } = 0;
+        public decimal InvestmentIncrease { get; set; } = 0;
     }
 }
